% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stg.R
\name{stl_dimensions}
\alias{stl_dimensions}
\title{Get dimensions of the object}
\usage{
stl_dimensions(stl)
}
\arguments{
\item{stl}{object from stl_load()}
}
\value{
matrix with min and max values for x, y, and z
}
\description{
Get dimensions of the object
}
