% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stg.R
\name{stl_load}
\alias{stl_load}
\title{Load STL file}
\usage{
stl_load(con)
}
\arguments{
\item{con}{A connection or filename}
}
\value{
A matrix with x, y, and z points
}
\description{
This is a wrapper over RGL's readSTL function
}
